refactor(zama-fhe): admit heap growth through HeapBudget
One running total gates Vec growth before the allocator
serves it. Purpose lists live on the stack. Frontier
proofs split under heap_budget/.
