+++ solana/crates/zama-fhe/src/lower.rs
@@ -1,14 +1,9 @@
 //! Lowers builder operands/outputs to the interned wire format.
 //!
-//! This file also owns the whole heap-tally protocol for a step: the three shared tables are
-//! [`TalliedVec`]s that pay for their own growth, and every exact-size allocation lowering
-//! causes (attestation embeds, subject index lists, purpose tables) goes through
-//! [`StepTables::tally_bytes`]. The protocol is *pre-emptive*: each site admits its bytes
-//! against the build budget via [`StepTables::ensure_headroom`] before the allocator serves
-//! them, so a step that gets rejected has only spent admitted bytes — the tally can never
-//! cross the budget, even transiently, on the never-freeing bump region. Nothing outside this
-//! file and `builder.rs` touches the tally — `accounts.rs` in particular is plain account
-//! metadata with no allocator accounting in it.
+//! Intern tables are [`TalliedVec`]s that grow through a [`HeapBudget`]. Exact-size allocations
+//! (attestation embeds, subject index lists) go through [`StepTables::tally_bytes`]. Purpose
+//! lists are a stack array and never touch the heap. On drop, an uncommitted step rolls back
+//! interned tables in place so a failed step leaves the builder as it found them.
 
 use zama_host::{CoprocessorInputAttestation, FheExecuteOperand, FheExecuteOutput};
 
@@ -16,7 +11,7 @@ use crate::accounts::{
     ExecutionAccountMeta, ExecutionAccountPurpose, ExecutionEncryptedValueAccountAuthority,
 };
 use crate::acl::{Output, OutputKind};
-use crate::heap_tally::{pushes_request, tally_push, TalliedVec};
+use crate::heap_tally::{HeapBudget, TalliedVec};
 use crate::operand::{Operand, OperandKind};
 use crate::{FheExecutionBuildError, Result};
 
@@ -42,26 +37,19 @@ pub(crate) struct StepTables<'b> {
     remaining_accounts: &'b mut TalliedVec<ExecutionAccountMeta>,
     dictionary: &'b mut TalliedVec<[u8; 32]>,
     persistent_producers: &'b mut TalliedVec<anchor_lang::prelude::Pubkey>,
-    /// The builder's explicit counter for exact-size allocations: attestation embeds, subject
-    /// index lists, purpose tables. Growth of the shared tables pays into the tables themselves;
-    /// rollback keeps every request — on the never-freeing bump region a rolled-back step's
-    /// requests are spent all the same.
-    explicit_bytes: &'b mut usize,
-    /// Requested bytes of the builder tables this step does not borrow (steps, produced types,
-    /// verified inputs) — frozen for the step's duration, so [`Self::ensure_headroom`] can see
-    /// the whole build's tally.
-    stable_bytes: usize,
+    budget: &'b mut HeapBudget,
     remaining_accounts_len: usize,
     dictionary_len: usize,
     persistent_producers_len: usize,
     promotions: TalliedVec<(usize, MetaPromotion)>,
+    committed: bool,
 }
 
 impl Drop for StepTables<'_> {
-    /// The undo log is step-local; its growth is builder cost like everything else, harvested
-    /// into the explicit counter whether the step committed or rolled back.
     fn drop(&mut self) {
-        *self.explicit_bytes += self.promotions.requested_bytes();
+        if !self.committed {
+            self.rollback();
+        }
     }
 }
 
@@ -70,8 +58,7 @@ impl<'b> StepTables<'b> {
         remaining_accounts: &'b mut TalliedVec<ExecutionAccountMeta>,
         dictionary: &'b mut TalliedVec<[u8; 32]>,
         persistent_producers: &'b mut TalliedVec<anchor_lang::prelude::Pubkey>,
-        explicit_bytes: &'b mut usize,
-        stable_bytes: usize,
+        budget: &'b mut HeapBudget,
     ) -> Self {
         Self {
             remaining_accounts_len: remaining_accounts.len(),
@@ -80,28 +67,24 @@ impl<'b> StepTables<'b> {
             remaining_accounts,
             dictionary,
             persistent_producers,
-            explicit_bytes,
-            stable_bytes,
+            budget,
             promotions: TalliedVec::new(),
+            committed: false,
         }
     }
 
-    /// Admits `upcoming` bytes against the build budget *before* the allocator serves them —
-    /// the pre-emptive half of the per-step heap gate. Every growing or exact-size allocation
-    /// below passes here first, so a rejected step has only spent admitted bytes and the
-    /// builder's tally can never cross the budget, even transiently.
-    fn ensure_headroom(&self, upcoming: usize) -> Result<()> {
-        if self.stable_bytes + self.requested_bytes() + upcoming
-            > crate::cost::BUILD_HEAP_BUDGET_BYTES
-        {
-            return Err(FheExecutionBuildError::ExceedsBuildHeapBudget);
-        }
-        Ok(())
+    pub(crate) fn budget(&mut self) -> &mut HeapBudget {
+        self.budget
+    }
+
+    /// Marks the step as kept. Drop then leaves interned tables as they are.
+    pub(crate) fn commit(&mut self) {
+        self.committed = true;
     }
 
     /// Undoes everything this step wrote: promotions newest-first, so an entry promoted twice ends
     /// on the oldest record, then the appended tails. Requested bytes stay requested.
-    pub(crate) fn rollback(&mut self) {
+    fn rollback(&mut self) {
         for index in (0..self.promotions.len()).rev() {
             let (meta_index, undo) = &self.promotions[index];
             let meta = self
@@ -121,33 +104,12 @@ impl<'b> StepTables<'b> {
     }
 
     pub(crate) fn account_index(&mut self, required: ExecutionAccountMeta) -> Result<u8> {
-        // Every constructed meta arrives with its one-purpose table already allocated (the few
-        // bytes of one purpose entry — the only allocation in the protocol that precedes its
-        // admission); charged here, kept, merged away, or rolled back.
-        let purposes_bytes =
-            required.purposes.len() * std::mem::size_of::<ExecutionAccountPurpose>();
-        self.ensure_headroom(purposes_bytes)?;
-        *self.explicit_bytes += purposes_bytes;
         if let Some(index) = self
             .remaining_accounts
             .iter()
             .position(|candidate| candidate.pubkey == required.pubkey)
         {
-            // Admit the merge's allocations before mutating anything: the purpose-table growth
-            // the appends will cause (simulated against the current table, skipping duplicates
-            // within `required` itself) plus the undo record.
             let meta = &self.remaining_accounts[index];
-            let appended_purposes = required
-                .purposes
-                .iter()
-                .enumerate()
-                .filter(|(position, purpose)| {
-                    !meta.purposes.contains(purpose)
-                        && !required.purposes[..*position].contains(purpose)
-                })
-                .count();
-            let merge_growth = pushes_request(&meta.purposes, appended_purposes);
-            self.ensure_headroom(merge_growth + self.promotions.next_push_request())?;
             let undo = MetaPromotion {
                 was_writable: meta.is_writable,
                 was_signer: meta.is_signer,
@@ -160,40 +122,21 @@ impl<'b> StepTables<'b> {
             meta.is_writable |= required.is_writable;
             meta.is_signer |= required.is_signer;
             for purpose in required.purposes {
-                if !meta.purposes.contains(&purpose) {
-                    tally_push(&meta.purposes, self.explicit_bytes);
-                    meta.purposes.push(purpose);
-                }
+                meta.purposes.try_insert(purpose);
             }
-            self.promotions.push((index, undo));
+            self.promotions.try_push(self.budget, (index, undo))?;
             return u8::try_from(index)
                 .map_err(|_| FheExecutionBuildError::TooManyRemainingAccounts);
         }
         let index = u8::try_from(self.remaining_accounts.len())
             .map_err(|_| FheExecutionBuildError::TooManyRemainingAccounts)?;
-        self.ensure_headroom(self.remaining_accounts.next_push_request())?;
-        self.remaining_accounts.push(required);
+        self.remaining_accounts.try_push(self.budget, required)?;
         Ok(index)
     }
 
-    /// Pays step-local allocation bytes into the builder's tally (per-step operand tables that
-    /// are not one of the three shared tables), admitting them against the budget first — call
-    /// before the allocation they price.
+    /// Pays an exact-size allocation into the budget before the allocator serves it.
     pub(crate) fn tally_bytes(&mut self, bytes: usize) -> Result<()> {
-        self.ensure_headroom(bytes)?;
-        *self.explicit_bytes += bytes;
-        Ok(())
-    }
-
-    /// Every byte requested so far by the parts of the builder this step borrows: the explicit
-    /// counter, the three shared tables, and the live undo log. `commit_step` adds the tables it
-    /// still holds and checks the per-step budget against the sum.
-    pub(crate) fn requested_bytes(&self) -> usize {
-        *self.explicit_bytes
-            + self.remaining_accounts.requested_bytes()
-            + self.dictionary.requested_bytes()
-            + self.persistent_producers.requested_bytes()
-            + self.promotions.requested_bytes()
+        self.budget.admit(bytes)
     }
 
     /// Interns a 32-byte constant into the execution dictionary, reusing an existing entry
@@ -205,8 +148,7 @@ impl<'b> StepTables<'b> {
         }
         let index = u8::try_from(self.dictionary.len())
             .map_err(|_| FheExecutionBuildError::TooManyDictionaryEntries)?;
-        self.ensure_headroom(self.dictionary.next_push_request())?;
-        self.dictionary.push(bytes);
+        self.dictionary.try_push(self.budget, bytes)?;
         Ok(index)
     }
 
@@ -218,9 +160,8 @@ impl<'b> StepTables<'b> {
         &mut self,
         encrypted_value: anchor_lang::prelude::Pubkey,
     ) -> Result<()> {
-        self.ensure_headroom(self.persistent_producers.next_push_request())?;
-        self.persistent_producers.push(encrypted_value);
-        Ok(())
+        self.persistent_producers
+            .try_push(self.budget, encrypted_value)
     }
 }
 
@@ -310,8 +251,7 @@ pub(crate) fn lower_output(
                     ))?)
                 };
             // One exact allocation for the index list, sized explicitly so the tally is exact.
-            tables.tally_bytes(binding.subjects.len() * std::mem::size_of::<u8>())?;
-            let mut output_subject_indexes = Vec::with_capacity(binding.subjects.len());
+            let mut output_subject_indexes = tables.budget().alloc_vec(binding.subjects.len())?;
             for subject in &binding.subjects {
                 output_subject_indexes.push(tables.dictionary_index(subject.to_bytes())?);
             }
